WIP: more tests

+++ api/src/main/java/io/mattinfern0/kanbanboardapi/tasks/TaskService.java
@@ -70,6 +70,9 @@ public class TaskService {
     }
 
     public void deleteTask(UUID taskId) {
+        if (!taskRepository.existsById(taskId)) {
+            throw new EntityNotFoundException(String.format("Task with id %s not found", taskId));
+        }
         taskRepository.deleteById(taskId);
     }
